requirements.txt
pymongo
flask
gunicorn
validators
python-dotenv